+++ src/tests_gui.py
@@ -7,6 +7,8 @@ GUI unit tests
 
 import unittest
 
+from test.utils import print_memory
+
 from test.guitests.test_mainwnd import MainWndTest
 from test.guitests.test_bookmarks import BookmarksGuiTest
 from test.guitests.test_attach import AttachPanelTest
@@ -50,21 +52,6 @@ from test.guitests.wikipage.test_wikitablerowsdialog import WikiTableRowsDialogT
 from test.guitests.wikipage.test_wikitableactions import WikiTableActionsTest
 
 
-def print_memory():
-    from pympler import muppy, summary
-    import gc
-    import wx
-
-    gc.collect()
-    all_objects = muppy.get_objects()
-    my_types = muppy.filter(all_objects, Type=wx.Object)
-    # my_types = muppy.filter(all_objects, Type=wx.Menu)
-    sum1 = summary.summarize(my_types)
-    summary.print_(sum1, limit=30)
-    # for t in my_types:
-    #     print(t)
-
-
 if __name__ == '__main__':
     unittest.main(exit=False)
-    # print_memory()
+    print_memory()
